Extract response processing into configurable classes
