+++ .rubocop.yml
@@ -74,9 +74,9 @@ Metrics/LineLength:
 Style/Documentation:
   Enabled: false
 
-Style/Lambda:
-  Exclude:
-    - 'lib/tlaw/response_processor.rb'
+# Style/Lambda:
+#   Exclude:
+#     - 'lib/tlaw/processors/response_processor.rb'
 
 # Those are just useless
 # ----------------------
