+++ lib/tlaw/processors.rb
@@ -1,0 +1,9 @@
+module TLAW
+  # TODO: Add docs
+  module Processors
+  end
+end
+
+require_relative 'processors/base'
+require_relative 'processors/data_table_response_processor'
+require_relative 'processors/response_processor'
